# Implement BinaryAppender