Implement the BinaryAppender interface in sketch types, so they can be marshalled to an existing slice.

Fixes https://github.com/axiomhq/hyperloglog/issues/46