@lukasmalkmus 2025-01-06
Hi @axw, thanks for the PR!

This looks promising so far, although I'm not the one to pull the trigger on this (@seiflotfy is), some general feedback:

1. Most of our Go OpenSource modules are compatible with the latest two major Go releases (and most likely even more if we're not going all-in on latest language/module additions). This PR would break that promise. My suggestion is to use build tags and keep the old implementation of `MarshalBinary` around for < Go 1.24. Anything >= Go 1.24, `MarshalBinary` can call `AppendBinary` under the hood.
2. Tests would be great but you pointed that out, already.
3. Normally I would ask for benchmarks. But looking at the code, it looks like there is no measurable impact to allocations? _Unless_ `AppendBinary` is called with a large enough buffer, allocations happen just like before with the exception that `compressedList.MarshalBinary/AppendBinary` now allocates only 8 bytes, initially? Correct me if I'm wrong.