@lukasmalkmus 2025-01-06
Ah, speaking tests, there is one thing I forgot: Ideally a test would also include an `Example` just to showcase potential memory savings via buffer reuse. Might even want to add comments around `MarshalBinary` pointing users to `AppendBinary`.