@axw 2025-01-06
> Most of our Go OpenSource modules are compatible with the latest two major Go releases (and most likely even more if we're not going all-in on latest language/module additions). This PR would break that promise. My suggestion is to use build tags and keep the old implementation of MarshalBinary around for < Go 1.24. Anything >= Go 1.24, MarshalBinary can call AppendBinary under the hood.

This shouldn't be a problem, as there's nothing referencing the `encoding.BinaryAppender` type. The types just happen to implement the new interface.

> Normally I would ask for benchmarks. But looking at the code, it looks like there is no measurable impact to allocations? Unless AppendBinary is called with a large enough buffer, allocations happen just like before with the exception that compressedList.MarshalBinary/AppendBinary now allocates only 8 bytes, initially? Correct me if I'm wrong.

That's correct.

> Ah, speaking tests, there is one thing I forgot: Ideally a test would also include an Example just to showcase potential memory savings via buffer reuse. Might even want to add comments around MarshalBinary pointing users to AppendBinary.

:+1: sounds good, will do